extract common transaction structs
