+++ minigu/storage/src/common/mod.rs
@@ -1,5 +1,6 @@
 pub mod iterators;
 pub mod model;
+pub mod transaction;
 pub mod wal;
 
 // Re-export commonly used types
@@ -7,4 +8,5 @@ pub use model::edge::*;
 pub use model::properties::*;
 pub use model::schema::*;
 pub use model::vertex::*;
+pub use transaction::*;
 pub use wal::*;
