+++ minigu/storage/src/tp_storage/checkpoint.rs
@@ -18,9 +18,9 @@ use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
 use super::memory_graph::{AdjacencyContainer, MemoryGraph, VersionedEdge, VersionedVertex};
-use super::transaction::Timestamp;
 use crate::common::model::edge::{Edge, Neighbor};
 use crate::common::model::vertex::Vertex;
+use crate::common::transaction::Timestamp;
 use crate::common::wal::StorageWal;
 use crate::common::wal::graph_wal::WalManagerConfig;
 use crate::error::{CheckpointError, StorageError, StorageResult};
